+++ facades/Response.d.ts
@@ -3,4 +3,5 @@ export default class Response {
     static setData(data?: any): ResponseBuilder;
     static setMessage(message: string): ResponseBuilder;
     static setStatus(status: number): ResponseBuilder;
+    static setCustom(custom?: Record<string, any>): ResponseBuilder;
 }